requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
scipy>=1.10.0

# Data processing
opencv-python>=4.7.0
scikit-image>=0.20.0
imageio>=2.28.0
face-alignment>=1.3.5
facenet-pytorch>=2.5.2

# Metrics and evaluation
pytorch-fid>=0.3.0
clean-fid>=0.1.35
lpips>=0.1.4
scikit-learn>=1.2.0

# Model architectures (optional - we'll implement from scratch)
timm>=0.9.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# GUI
gradio>=3.35.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Latent editing
ninja>=1.11.0  # For custom CUDA ops

# Ethical AI
fairface  # Bias detection
imagehash>=4.3.1  # Duplicate detection

# Optional cloud storage
boto3>=1.26.0  # AWS S3
google-cloud-storage>=2.9.0  # GCS